refactor: extract shared resume candidate preparation from route files into service module (#961)
Eliminates 725 lines of duplicated code between resumes.ts and
resumes_search.ts by extracting shared types, helpers, and the large
prepareConvexCandidates function (~260 lines, previously duplicated
verbatim) into services/resume-candidate-prep.ts.
